server/requirements.txt
Flask==3.1.1
flask_cors==6.0.0
flask_sqlalchemy==3.1.1
PyJWT==2.10.1
PyJWT==2.10.1
python_bcrypt==0.3.2
